refactor(security): move the PII log helper out of a heavy package
backend/utils/__init__.py eagerly imports AsyncHttpService, the message
chunker and the query builder, so `from backend.utils.pii_log_identifier
import ...` drags app settings in behind it. That broke a real contract:
scripts/intake_reprocess_backlog.py must answer --help without app
settings, and the moment crm_push.py imported the helper it could not.

The helper itself is clean — hmac, os, re, hashlib, nothing else. The
defect was where it was put, not what it does. backend/security/ is a
one-line docstring package that already hosts webhook_verifier.py, which
is exactly what a redaction primitive is: a security primitive with no
import weight, importable from anywhere including a bare CLI script.

Caught by the full backend suite, not by the tests for the code that
moved. A log-redaction helper is imported by things that deliberately
avoid the app — that is precisely why it must stay light.
