+++ apps/backend-rag/backend/agents/services/whatsapp_notification.py
@@ -9,7 +9,7 @@ from __future__ import annotations
 import asyncio
 import logging
 
-from backend.utils.pii_log_identifier import redact_identifier_for_log
+from backend.security.pii_log_identifier import redact_identifier_for_log
 
 logger = logging.getLogger(__name__)
