+++ apps/backend-rag/backend/app/routers/twitter.py
@@ -9,7 +9,7 @@ from fastapi import APIRouter, Request
 from fastapi.responses import JSONResponse
 
 from backend.app.core.config import settings
-from backend.utils.pii_log_identifier import redact_identifier_for_log
+from backend.security.pii_log_identifier import redact_identifier_for_log
 
 logger = logging.getLogger(__name__)
 router = APIRouter(prefix="/api/twitter", tags=["twitter"])
