+++ apps/backend-rag/backend/channels/instagram/adapter.py
@@ -9,7 +9,7 @@ import httpx
 from backend.channels.base import BaseChannel, ChannelMessage, ChannelResponse
 from backend.channels.instagram.config import InstagramChannelConfig
 from backend.channels.instagram.formatter import InstagramMessageFormatter
-from backend.utils.pii_log_identifier import redact_identifier_for_log
+from backend.security.pii_log_identifier import redact_identifier_for_log
 
 logger = logging.getLogger(__name__)
