+++ apps/backend-rag/backend/channels/optimizations.py
@@ -24,7 +24,7 @@ from datetime import datetime, timedelta, timezone
 from typing import Any
 
 from backend.core.secret_log_redaction import install_telegram_token_redaction
-from backend.utils.pii_log_identifier import redact_identifier_for_log
+from backend.security.pii_log_identifier import redact_identifier_for_log
 
 install_telegram_token_redaction()
