+++ apps/backend-rag/backend/channels/telegram/adapter.py
@@ -15,7 +15,7 @@ from backend.channels.base import BaseChannel, ChannelMessage, ChannelResponse
 from backend.channels.telegram.config import TelegramChannelConfig
 from backend.channels.telegram.formatter import TelegramMessageFormatter
 from backend.services.integrations.telegram_bot_service import TelegramBotService
-from backend.utils.pii_log_identifier import redact_identifier_for_log
+from backend.security.pii_log_identifier import redact_identifier_for_log
 
 logger = logging.getLogger(__name__)
