+++ apps/backend-rag/backend/services/intake/crm_push.py
@@ -36,7 +36,7 @@ from typing import Any
 import httpx
 
 from backend.phone_lock import phone_core
-from backend.utils.pii_log_identifier import redact_identifier_for_log
+from backend.security.pii_log_identifier import redact_identifier_for_log
 
 logger = logging.getLogger(__name__)
