+++ apps/backend-rag/backend/services/integrations/human_escalation_notifier.py
@@ -49,7 +49,7 @@ import time
 
 from backend.app.core.config import settings
 from backend.services.integrations.telegram_bot_service import telegram_bot
-from backend.utils.pii_log_identifier import redact_identifier_for_log
+from backend.security.pii_log_identifier import redact_identifier_for_log
 
 logger = logging.getLogger(__name__)
