+++ apps/backend-rag/backend/services/integrations/instagram_service.py
@@ -15,7 +15,7 @@ import httpx
 
 from backend.app.core.config import settings
 from backend.app.core.constants import HttpTimeoutConstants
-from backend.utils.pii_log_identifier import redact_identifier_for_log
+from backend.security.pii_log_identifier import redact_identifier_for_log
 
 logger = logging.getLogger(__name__)
