+++ apps/backend-rag/backend/services/integrations/messaging_identity_service.py
@@ -26,7 +26,7 @@ from typing import Any
 
 import asyncpg
 
-from backend.utils.pii_log_identifier import redact_identifier_for_log
+from backend.security.pii_log_identifier import redact_identifier_for_log
 
 logger = logging.getLogger(__name__)
