+++ apps/backend-rag/backend/services/integrations/whatsapp_triage_service.py
@@ -13,7 +13,7 @@ import logging
 from enum import Enum
 
 from backend.app.core.config import settings
-from backend.utils.pii_log_identifier import redact_identifier_for_log
+from backend.security.pii_log_identifier import redact_identifier_for_log
 
 logger = logging.getLogger(__name__)
