+++ apps/backend-rag/backend/services/whatsapp_context_builder.py
@@ -18,7 +18,7 @@ from typing import Any
 
 import asyncpg
 
-from backend.utils.pii_log_identifier import redact_identifier_for_log
+from backend.security.pii_log_identifier import redact_identifier_for_log
 
 logger = logging.getLogger(__name__)
