+++ apps/backend-rag/backend/services/whatsapp_onboarding_detector.py
@@ -14,7 +14,7 @@ from typing import Any
 
 import httpx
 
-from backend.utils.pii_log_identifier import redact_identifier_for_log
+from backend.security.pii_log_identifier import redact_identifier_for_log
 
 logger = logging.getLogger(__name__)
